Backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-jose
passlib[bcrypt]
python-multipart